fix: arrumando packages e s3 funcionando (#55)
