+++ log/banco-de-dados.trace.db
@@ -46,7 +46,7 @@ DROP TABLE IF EXISTS empresa [90107-224]
 	at org.springframework.jdbc.core.JdbcTemplate$1ExecuteStatementCallback.doInStatement(JdbcTemplate.java:435)
 	at org.springframework.jdbc.core.JdbcTemplate.execute(JdbcTemplate.java:393)
 	at org.springframework.jdbc.core.JdbcTemplate.execute(JdbcTemplate.java:444)
-	at sptech.fynx.dao.sptech.fynx.Main.main(sptech.fynx.Main.java:13)
+	at sptech.fynx.model.dao.sptech.fynx.Main.main(sptech.fynx.Main.java:13)
 2025-03-30 22:11:27.008352-03:00 jdbc[3]: exception
 org.h2.jdbc.JdbcSQLSyntaxErrorException: Syntax error in SQL statement "CREATE TABLE IF NOT EXISTS login (\000a\0009idLogin INT AUTO_INCREMENT,\000a\0009fkFuncionario INT,\000a\0009fkEmpresa INT,\000a\0009gerente BOOLEAN,\000a\0009email VARCHAR(100),\000a\0009senha VARCHAR(50),\000a\000a\0009CONSTRAINT [*]PRIMARY KEY (idLogin, fkFuncionario, fkEmpresa),\000a\000a\0009CONSTRAINT fk_login_funcionario\000a\0009FOREIGN KEY (fkFuncionario)\000a\0009\0009REFERENCES funcionario(idFuncionario)\000a);"; expected "identifier"; SQL statement:
 CREATE TABLE IF NOT EXISTS login (
