+++ log/src/main/java/sptech/fynx/LogLocal.java
@@ -6,7 +6,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.ThreadLocalRandom;
 
-public class Log {
+public class LogLocal {
 
     static List<String> logs = new ArrayList<>();
