+++ log/src/main/java/sptech/fynx/model/dao/ConexaoBD.java
@@ -1,4 +1,4 @@
-package sptech.fynx.dao;
+package sptech.fynx.model.dao;
 
 import org.springframework.jdbc.datasource.DriverManagerDataSource;
